Terraform for CICD (CodePipleline & CodeBuild)
We want to start any new project with automation. So to kick things off, a shell of CICD based on CodePipeline and CodeBuild is deployed with Terraform. Most deployment changes can be handled within buildspec.yaml files for code/web, infrastructure and deploy, leaving this initial automation largely untouched.

Also
+ .gitignore - A starting point generated from https://gitignore.io
+ docs updated