Hosting Infrastructure (#6)
Updates to CICD to provide needed information to Infrastructure code

Full hosting infrastructure built with Terraform, with code automatically deployed by CodeBuild. SSL and CDN are in place, as is a redirect from the www subdomain to the apex.

This is our starting point for a complete CICD solution that pushes code to the web. Much more can be done, but this PR generates a working web site.